being refactor to makeCodec
