+++ packages/fireproof/src/eb-node.ts
@@ -1,0 +1,4 @@
+import { EncryptedBlockstore, TransactionMeta } from '@fireproof/encrypted-blockstore'
+
+
+//
